DA preparation: obsope_parallel bug and obs output bug fixed
FossilOrigin-Name: 50d12f1cae966842964985845f004ddf7f0664757b0f57157a192617802ff052

+++ MSM-Tactical/sys/ush/rensda.sh
@@ -5,29 +5,32 @@ cycleda=${2}
 head_bv=${HEAD:-bv}
 head_da=${HEAD2:-da}
 ## experiment parameters
+# ensemble size
+member=${MEMBER:-10}
+# analysis date
+adate=${SDATE:-2022061812}
+# first guess lead time
+fhour=${ENDHOUR:-0}
+# model resolution
 ires=${IRES:-27}
+# observation settings
 single=${SINGLEOBS:-F}
-if [ $single = T ];then
-  debug_obs=T
-  print_img=5
-fi
 lats=${DA_SINGLE_LATS}
 latn=${DA_SINGLE_LATN}
 lonw=${DA_SINGLE_LONW}
 lone=${DA_SINGLE_LONE}
 fixed_level=${DA_FIXED_LEVEL}
+lmin=${LOBSMIN:--60}
+rmin=${ROBSMIN:-60}
+selobs=${SELOBS:-all}
+prep=${PREP:-_preprh}
+# parallelization
 nisep=${NISEP:-5}
 njsep=${NJSEP:-2}
 ighost=${IGHOST:-1}
 jghost=${JGHOST:-1}
 NODE=`expr $nisep \* $njsep`
-member=${MEMBER:-10}
-adate=${SDATE:-2022061812}
-fhour=${ENDHOUR:-0}
-lmin=${LOBSMIN:--60}
-rmin=${ROBSMIN:-60}
-selobs=${SELOBS:-all}
-prep=${PREP:-_preprh}
+# DA parameters
 mean=${DA_MEAN}
 tl=${DA_TL}
 scl_mem=${DA_SCL_MEM}
@@ -35,7 +38,9 @@ maxiter=${MAXITER:-5}
 hloc=${HLOC}
 vloc=${VLOC}
 minfl=${MINFL}
-q_update_top=${Q_UPDATE_TOP}
+rtpp=${RTPP}
+rtps=${RTPS}
+relax_spread_out=${RELAX_SPREAD_OUT}
 if [ ! -z $hloc ];then
 sigh=${hloc}.0d3
 fi
@@ -44,6 +49,21 @@ sigv=${vloc}.0d-1
 fi
 if [ ! -z $minfl ];then
 infl_mul=${minfl}.0d-2
+fi
+if [ ! -z $rtpp ];then
+infl_rtpp=${rtpp}.0d-2
+fi
+if [ ! -z $rtps ];then
+infl_rtps=${rtps}.0d-2
+fi
+q_update_top=${Q_UPDATE_TOP}
+save_info=${DA_SAVE_INFO}
+## end of inputs
+#debug_obs=T #debug
+print_img=$NODE #debug
+if [ $single = T ];then
+  debug_obs=T
+  print_img=5
 fi
 #saveens=${SAVEENS:-0} #0:save all ensemble, 1:save only ctrl, mean and spread
 ## data directories
@@ -173,7 +193,15 @@ cat <<EOF >lmlef.nml
  sigma_obsv=${sigv},
  gross_error=,
  cov_infl_mul=${infl_mul},
+ sp_infl_rtpp=${infl_rtpp},
+ sp_infl_rtps=${infl_rtps},
+ relax_spread_out=${relax_spread_out},
  maxiter=${maxiter},
+ nonlinear=,
+ zupd=,
+ save_info=${save_info},
+ info_out_basename=,
+ ewgt_basename=,
  q_update_top=${q_update_top},
  q_adjust=T,
  oma_monit=T,
@@ -187,7 +215,13 @@ EOF
 rm -f ${obsf}.dat
 ln -s ${obsdir}/${adate}/${obsf}.dat .
 rm -f gues.*.grd ${obsinf}.*.dat ${obsextf}.*.dat
-ln -s ${guesdir}/${pdate}/r_sig.f$fh gues.0000.grd
+if [ $cycleda -eq 1 ];then
+ln -s ${guesdir}/${pdate}/r_sig.f$fh gues.0000.sig.grd
+ln -s ${guesdir}/${pdate}/r_sfc.f$fh gues.0000.sfc.grd
+else
+ln -s ${guesdir}/${pdate}/${head}000/r_sig.f$fh gues.0000.sig.grd
+ln -s ${guesdir}/${pdate}/${head}000/r_sfc.f$fh gues.0000.sfc.grd
+fi
 if [ $obsda_in = T ];then
 ln -s ${obsdir}/${pdate}/${obsextf}.0000.dat .
 fi
@@ -197,7 +231,8 @@ fi
 m=1
 while [ $m -le $member ];do
 mem=`printf '%0.3d' $m`
-ln -s ${guesdir}/${pdate}/${head}${mem}/r_sig.f$fh gues.0${mem}.grd
+ln -s ${guesdir}/${pdate}/${head}${mem}/r_sig.f$fh gues.0${mem}.sig.grd
+ln -s ${guesdir}/${pdate}/${head}${mem}/r_sfc.f$fh gues.0${mem}.sfc.grd
 if [ $obsda_in = T ];then
 ln -s ${obsdir}/${adate}/${obsextf}.0${mem}.dat .
 fi
@@ -210,11 +245,11 @@ rm -f ${obsoutf}.*.dat anal.*.grd
 mpiexec -n $NODE ./lmlef 2>${logf}err || exit 11
 if [ ! -z ${print_img} ];then
 img=`printf '%0.3d' $print_img`
-mv NOUT-$img ${logf}n${NODE}.txt
+mv NOUT-$img ${logf}txt
 else
-mv NOUT-001 ${logf}n${NODE}.txt
+mv NOUT-001 ${logf}txt
 fi
-rm -f NOUT-*
+#rm -f NOUT-*
 #fi # ! -f ${logf}_n${NODE}.txt
 
 ## post process
@@ -222,9 +257,9 @@ rm -f NOUT-*
 for vtype in gues anal;do
 for emem in ctrl mean sprd;do
 if [ $emem = ctrl ];then
-in=${vtype}.0000.grd
+in=${vtype}.0000.sig.grd
 else
-in=${vtype}.${emem}.grd
+in=${vtype}.${emem}.sig.grd
 fi
 out=${vtype}.${emem}.bin
 ctl=${vtype}.${emem}.ctl
@@ -244,7 +279,67 @@ sed -i -e 's/DATAFILE/'$out'/g' $ctl
 rm ${ctl}-e
 done
 done
+if [ $relax_spread_out = T ];then
+  rm -f fort.*
+  ln -s rtps.sig.grd fort.11
+  ln -s rtps.bin fort.51
+  ln -s rtps.ctl fort.61
+  echo "rtps" >> read_sig.log
+  ./read_sig < read_sig.nml >> read_sig.log || exit 11
+  sed -i -e 's/DATAFILE/'rtps.bin'/g' rtps.ctl
+  rm rtps.ctl-e
+fi
+if [ $save_info = T ];then
+cat << EOF > read_info.nml
+&namlst_info
+ member=${member},
+ emem=F,
+ writectl=T,
+&end
+EOF
+  rm -f fort.* read_info
+  ln -s dainfo.sig.grd fort.11
+  ln -s dainfo.bin fort.51
+  ln -s dainfo.ctl fort.61
+  ln -s ${bindir}/read_info .
+  ./read_info < read_info.nml || exit 12
+  sed -i -e 's/DATAFILE/dainfo.bin/g' dainfo.ctl
+  rm dainfo.ctl-e
+  m=1
+  while [ $m -le $member ];do
+  mem=`printf '%0.3d' $m`
+  rm fort.*
+  if [ $m -eq 1 ];then
+cat << EOF > read_info.nml
+&namlst_info
+ member=${member},
+ emem=T,
+ writectl=T,
+&end
+EOF
+  ln -s ewgt.ctl fort.61
+else
+cat << EOF > read_info.nml
+&namlst_info
+ member=${member},
+ emem=T,
+ writectl=F,
+&end
+EOF
+  fi
+  ln -s ewgt.0${mem}.sig.grd fort.11
+  ln -s ewgt.0${mem}.bin fort.51
+  ./read_info < read_info.nml || exit 12
+  if [ $m -eq 1 ];then
+  sed -i -e 's/DATAFILE/..\/'${head_da}'%e\/ewgt.0%e.bin/g' ewgt.ctl
+  rm ewgt.ctl-e
+  fi
+  m=`expr $m + 1`
+  done
+fi
 ### prepare forecast initial files
+mkdir -p $wdir/${head_da}000
+cd $wdir/${head_da}000
 ### copy namelists
 cp $guesdir/$pdate/rsmparm .
 cp $guesdir/$pdate/rsmlocation .
@@ -255,18 +350,30 @@ cp $guesdir/$pdate/rmtn.parm .
 cp $guesdir/$pdate/rmtnoss .
 cp $guesdir/$pdate/rmtnslm .
 cp $guesdir/$pdate/rmtnvar .
-mkdir -p $wdir/${head_da}000
-cd $wdir/${head_da}000
-mv ../anal.0000.grd r_sig.f00
+mv ../anal.0000.sig.grd r_sig.f00
+mv ../anal.0000.sfc.grd r_sfc.f00
+mv ../gues.ctrl.bin .
+mv ../gues.ctrl.ctl .
 mv ../anal.ctrl.bin .
 mv ../anal.ctrl.ctl .
+if [ -f ../${obsinf}.0000.dat ]; then
 mv ../${obsinf}.0000.dat .
+fi
+if [ -f ../${obsoutf}.0000.dat ]; then
 mv ../${obsoutf}.0000.dat . 
-cd $wdir
-cp $wdir/${head_da}000/r_sig.f00 .
+fi
+if [ $relax_spread_out = T ];then
+mv ../rtps.* .
+fi
+if [ $save_info = T ];then
+mv ../dainfo.* .
+mv ../ewgt.ctl .
+fi
+#cd $wdir
+#cp $wdir/${head_da}000/r_sig.f00 .
+#cp $wdir/${head_da}000/r_sfc.f00 .
 cp r_sig.f00 r_sigi
 cp r_sig.f00 r_sigitdt
-cp ${guesdir}/${pdate}/r_sfc.f$fh r_sfc.f00
 cp r_sfc.f00 r_sfci
 m=1
 while [ $m -le $member ];do
@@ -283,60 +390,35 @@ cp $guesdir/$pdate/${head}${mem}/rmtn.parm .
 cp $guesdir/$pdate/${head}${mem}/rmtnoss .
 cp $guesdir/$pdate/${head}${mem}/rmtnslm .
 cp $guesdir/$pdate/${head}${mem}/rmtnvar .
-mv ../anal.0$mem.grd r_sig.f00
+mv ../anal.0$mem.sig.grd r_sig.f00
+mv ../anal.0$mem.sfc.grd r_sfc.f00
+if [ -f ../${obsinf}.0$mem.dat ]; then
 mv ../${obsinf}.0$mem.dat .
-mv ../${obsoutf}.0$mem.dat . 
-cp ${guesdir}/${pdate}/${head}${mem}/r_sfc.f$fh r_sfc.f00
+fi
+if [ -f ../${obsoutf}.0$mem.dat ]; then
+mv ../${obsoutf}.0$mem.dat .
+fi
+if [ $save_info = T ]; then
+mv ../ewgt.0$mem.* .
+fi
 cp r_sig.f00 r_sigi
 cp r_sig.f00 r_sigitdt
 cp r_sfc.f00 r_sfci
-cd -
+cd $wdir
 m=`expr $m + 1`
 done
 for emem in mean sprd;do
 mkdir -p ${guesdir}/${pdate}/${head}${emem}
-mv gues.${emem}.grd ${guesdir}/${pdate}/${head}${emem}/r_sig.f$fh
+mv gues.${emem}.sig.grd ${guesdir}/${pdate}/${head}${emem}/r_sig.f$fh
+mv gues.${emem}.sfc.grd ${guesdir}/${pdate}/${head}${emem}/r_sfc.f$fh
 mv gues.${emem}.bin ${guesdir}/${pdate}/${head}${emem}/
 mv gues.${emem}.ctl ${guesdir}/${pdate}/${head}${emem}/
 mkdir -p ${wdir}/${head_da}${emem}
-mv anal.${emem}.grd ${wdir}/${head_da}${emem}/r_sig.f00
+mv anal.${emem}.sig.grd ${wdir}/${head_da}${emem}/r_sig.f00
+mv anal.${emem}.sfc.grd ${wdir}/${head_da}${emem}/r_sfc.f00
 mv anal.${emem}.bin ${wdir}/${head_da}${emem}/
 mv anal.${emem}.ctl ${wdir}/${head_da}${emem}/
 done
-#if [ $saveens -eq 1 ];then
-#m=1
-#while [ $m -le $member ];do
-#mem=`printf '%0.4d' $m`
-#rm ${obsoutf}.${mem}.dat
-#rm ${anloutf}.${mem}.grd
-#m=`expr $m + 1`
-#done
-#fi
-#
-#
-#if [ $saveens -eq 0 ];then
-#m=1
-#while [ $m -le $member ];do
-#mem=`printf '%0.4d' $m`
-#in=${anloutf}.${mem}.grd
-#out=${anloutf}.${mem}.bin
-#ctl=${anloutf}.${mem}.ctl
-#rm -f fort.* read_sig
-#ln -s ${in} fort.11
-#ln -s ${out} fort.51
-#ln -s ${ctl} fort.61
-#ln -s ${bindir2}/read_sig .
-#cat << EOF > read_sig.nml
-#&namlst_cld
-# icld=0,
-#&end
-#EOF
-#./read_sig < read_sig.nml
-#sed -i -e 's/DATAFILE/'$out'/g' $ctl
-#rm ${ctl}-e
-#m=`expr $m + 1`
-#done
-#fi
 rm gues.*.grd
 if [ $obsda_in = T ];then
 rm ${obsinf}.*.dat
